check_lang script can correct encoding

+++ trunk/scripts/getlang.sh
@@ -10,7 +10,13 @@ echo Directory $2 is in my way!
 fi
 
 cp -R $1 $2
-cd $2
-find . -type f ! \(  -name "$2.js" -o  -name "$2.inc.php" \) -exec rm -f {} \;
-find -depth -type d -empty -exec rmdir {} \;
+
+if [ -d $2 ]; then
+	cd $2
+	find . -type f ! \(  -name "$2.js" -o  -name "$2.inc.php" \) -exec rm -f {} \;
+	find -depth -type d -empty -exec rmdir {} \;
+else
+	echo "Error: could not create $2";
+	exit
+fi
